#3013 delete new utils file
