+++ wls-gui-wahllokalsystem/src/components/dse/BaseTeamStatusListItem.vue
@@ -46,8 +46,8 @@ import type { StimmzettelerfassungTeamStatusEntry } from "@/types/dse/Stimmzette
 import type { PropType } from "vue";
 
 import BaseTextButton from "@/components/common/buttons/BaseTextButton.vue";
-import { useDseWorkflowStatusUtils } from "@/composables/dse/dseWorkflowStatusUtils.ts";
 import { useMonitoringViewButtonStateUtils } from "@/composables/dse/monitoringViewButtonStateUtils.ts";
+import { useMonitoringViewUtils } from "@/composables/dse/monitoringViewUtils.ts";
 import { useStimmzettelerfassungTeamStatusMapper } from "@/composables/dse/stimmzettelerfassungTeamStatusMapper.ts";
 import { useStimmzettelErfassungViewUtils } from "@/composables/dse/stimmzettelErfassungViewUtils.ts";
 
@@ -82,7 +82,7 @@ const { sendStatusInBearbeitung } = useStimmzettelErfassungViewUtils(
   props.wahlbezirkID,
   props.teamEntry.teamID
 );
-const { workflowStatus } = useDseWorkflowStatusUtils(
+const { workflowStatus } = useMonitoringViewUtils(
   props.wahlID,
   props.wahlbezirkID
 );
