+++ wls-gui-wahllokalsystem/src/views/dse/MonitoringView.vue
@@ -85,7 +85,7 @@ import BaseLatestLoadDiv from "@/components/common/div/BaseLatestLoadDiv.vue";
 import BaseProgressLinear from "@/components/common/progressLinear/BaseProgressLinear.vue";
 import BaseTeamStatusListItem from "@/components/dse/BaseTeamStatusListItem.vue";
 import TheBeschlussfassungStartenDialog from "@/components/dse/TheBeschlussfassungStartenDialog.vue";
-import { useDseWorkflowStatusUtils } from "@/composables/dse/dseWorkflowStatusUtils.ts";
+import { useDseWorkflowStatusService } from "@/composables/dse/dseWorkflowStatusService.ts";
 import { useMonitoringViewUtils } from "@/composables/dse/monitoringViewUtils.ts";
 import { StimmzettelerfassungStatusEnum } from "@/types/dse/StimmzettelerfassungStatusEnum.ts";
 import { StimmzettelerfassungTeamStatusEnum } from "@/types/dse/StimmzettelerfassungTeamStatusEnum.ts";
@@ -104,10 +104,7 @@ const {
   workflowStatus,
   onMonitoringSynchronisierenClicked,
 } = useMonitoringViewUtils(wahlID, wahlbezirkID);
-const { reloadWorkflowStatus } = useDseWorkflowStatusUtils(
-  wahlID,
-  wahlbezirkID
-);
+const { loadDseWorkflowStatus } = useDseWorkflowStatusService();
 
 const beschlussfassungStartenBtnActive = computed(() =>
   teamstatusList.value.every(
@@ -136,7 +133,7 @@ async function onBeschlussfassungStartenClicked() {
   beschlussfassungStartenDialogVisible.value = true;
 }
 async function onAktualisierenClicked() {
-  await reloadWorkflowStatus();
+  await loadDseWorkflowStatus(wahlID, wahlbezirkID, false);
   await onMonitoringSynchronisierenClicked(true);
 }
 </script>
